requirements.txt
pandas==1.3.5
psycopg2==2.9.1
numpy==1.21.4
python-dotenv==0.19.1
openpyxl==3.0.9